requirements.txt
streamlit>=1.45.0
google-genai
huggingface-hub
rdkit-pypi
biopython
Pillow
requests
pandas
numpy
matplotlib
plotly
